Move DefinitionResult to new directory.
