+++ src/components/pages/GameContainer/Game/DefinitionResult/index.ts
@@ -1,0 +1,1 @@
+export { default as DefinitionResult } from './DefinitionResult';
