+++ src/components/pages/GameContainer/Game/Postgame/Postgame.tsx
@@ -27,7 +27,7 @@ import { Player } from '../../../../common/Player';
 import { PlayerList } from '../../../../common/PlayerList';
 import { ProTip } from '../../../../common/ProTip';
 import { SetHost } from '../../../../common/SetHost';
-import { DefinitionResult } from './DefinitionResult';
+import { DefinitionResult } from '../DefinitionResult/';
 
 const Postgame = (props: PostgameProps): React.ReactElement => {
   const {
